requirements.txt
httpx==0.27.0
PyYAML==6.0.2
pydantic==2.8.2
jmespath==1.0.1 